refacto delete tests + rename lightning_refacto1 -> autoregressive
