+++ bin/lightning_cli/cli_autoregressive.py
@@ -1,7 +1,6 @@
 from lightning.pytorch.cli import LightningCLI, ArgsType
-from py4cast.lightning_module.lightning_refacto1 import AutoRegressiveLightningModule
+from py4cast.lightning_module.autoregressive import AutoRegressiveLightningModule
 from py4cast.data_module.dummy import DummyDataModule
-
 
 class MyLightningCLI(LightningCLI):
     def add_arguments_to_parser(self, parser):
