+++ bin/lightning_cli/cli_mae.py
@@ -3,8 +3,19 @@ from py4cast.lightning_module.mae import MAELightningModule
 from py4cast.data_module.dummy import DummyDataModule
 
 
+class MyLightningCLI(LightningCLI):
+    def add_arguments_to_parser(self, parser):
+        parser.link_arguments(
+            "data.train_dataset_info", "model.dataset_info", apply_on="instantiate"
+        )
+        parser.link_arguments("data.dataset_name", "model.dataset_name")
+        parser.link_arguments(
+            "data.batch_shape", "model.batch_shape", apply_on="instantiate"
+        )
+
+
 def cli_main(args: ArgsType = None):
-    cli = LightningCLI(
+    cli = MyLightningCLI(
         MAELightningModule,
         DummyDataModule,
         run=False,
@@ -18,6 +29,5 @@ if __name__ == "__main__":
     cli_main(
         [
             "--config=config/lightning_cli/config_cli_mae.yaml",
-            "--trainer.fast_dev_run=False",
         ]
     )
