+++ py4cast/lightning_module/utils.py
@@ -119,6 +119,7 @@ class PlotLightningModule:
             dataformats="HW",
         )
 
+
 ###--------------------- MASK ---------------------###
